refractor (Vendors) : manage shop allow vendors to manage their own shops

+++ products/views.py
@@ -1,7 +1,8 @@
 from django.urls import reverse
 from orders.models import Order
+from vendors.models import Vendor
 from django.contrib import messages
-from products.decorators import is_admin
+from products.decorators import is_vendor
 from customers.decorators import is_customer
 from products.models import Product,Price,Content
 from django.contrib.auth.decorators import login_required
@@ -11,11 +12,11 @@ from django.http import Http404,HttpResponse,HttpResponseRedirect
 def get_product(pk):
     return get_object_or_404(Product,pk=pk)
 
-def get_content(pk):
-    return get_object_or_404(Content,pk=pk)
-
 def get_price(pk):
     return get_object_or_404(Price,pk=pk)
+
+def get_content(pk):
+    return get_object_or_404(Content,pk=pk)
 
 def set_product_details(request,product):
     form_contents= request.POST.getlist("content-input")
@@ -36,8 +37,7 @@ def set_product_details(request,product):
         )
         price.save()
 
-def get_product_catalogue():
-    products= Product.objects.all()
+def get_product_catalogue(products):
     catalogue= dict()
     for product in products:
         contents= Content.objects.filter(product=product)
@@ -56,85 +56,109 @@ def get_product_catalogue():
     return catalogue
 
 @login_required(login_url="user:login")
-@is_admin
-def all_products(request):
-    context= {"catalogue":get_product_catalogue()}
-    return render(request,"products/all_products.html",context)
+@is_vendor
+def my_shop(request):
+    vendor= Vendor.objects.get(id=request.user.id)
+    products= Product.objects.filter(vendor=vendor)
+    context= {"catalogue":get_product_catalogue(products)}
+    return render(request,"products/my_shop.html",context)
 
 @login_required(login_url="user:login")
-@is_admin
+@is_vendor
 def add_product(request):
-    context= {}
     if request.method == 'POST':
-        form_product= request.POST.get("product-name")
-        product= Product.objects.create(name=form_product)
+        vendor= Vendor.objects.get(id=request.user.id)
+        product_name= request.POST.get("product-name")
+        product= Product.objects.create(
+            vendor=vendor,
+            name=product_name,
+        )
         product.save()
         set_product_details(request,product)
-        return HttpResponseRedirect(reverse("products:all-products"))
+        return HttpResponseRedirect(reverse("products:my-shop"))
+    context= {}
     return render(request,"products/add_product.html",context)
 
 @login_required(login_url="user:login")
-@is_admin
+@is_vendor
 def edit_product(request,pk):
+    vendor= Vendor.objects.get(id=request.user.id)
     product= get_product(pk)
+    if vendor != product.vendor:
+        return HttpResponse("You are not allowed to access this page")
     if request.method == "POST":
         product.name= request.POST.get("edit-product-name")
         product.save()
         set_product_details(request,product)
-        return HttpResponseRedirect(reverse("products:all-products"))
+        return HttpResponseRedirect(reverse("products:my-shop"))
     context= {"product":product}
     return render(request,"products/edit_product.html",context)
 
 @login_required(login_url="user:login")
-@is_admin
+@is_vendor
 def delete_product(request,pk):
+    vendor= Vendor.objects.get(id=request.user.id)
     product= get_product(pk)
+    if vendor != product.vendor:
+        return HttpResponse("You are not allowed to access this page")
     if request.method == "POST":
         product.delete()
-        return HttpResponseRedirect(reverse("products:all-products"))
+        return HttpResponseRedirect(reverse("products:my-shop"))
     context= {"obj":product}
     return render(request,"delete.html",context)
 
 @login_required(login_url="user:login")
-@is_admin
+@is_vendor
 def edit_content(request,pk):
+    vendor= Vendor.objects.get(id=request.user.id)
     content= get_content(pk)
+    if vendor != content.product.vendor:
+        return HttpResponse("You are not allowed to access this page")
     if request.method == "POST":
         content.name= request.POST.get("edit-content-name")
         content.save()
-        return HttpResponseRedirect(reverse("products:all-products"))
+        return HttpResponseRedirect(reverse("products:my-shop"))
     context= {"content":content}
     return render(request,"products/edit_content.html",context)
 
 @login_required(login_url="user:login")
-@is_admin
+@is_vendor
 def delete_content(request,pk):
+    vendor= Vendor.objects.get(id=request.user.id)
     content= get_content(pk)
+    if vendor != content.product.vendor:
+        return HttpResponse("You are not allowed to access this page")
     if request.method == "POST":
         content.delete()
-        return HttpResponseRedirect(reverse("products:all-products"))
+        return HttpResponseRedirect(reverse("products:my-shop"))
     context= {"obj":content}
     return render(request,"delete.html",context)
 
 @login_required(login_url="user:login")
-@is_admin
+@is_vendor
 def edit_price(request,pk):
+    vendor= Vendor.objects.get(id=request.user.id)
     price= get_price(pk)
+    if vendor != price.product.vendor:
+        return HttpResponse("You are not allowed to access this page")
     if request.method == "POST":
         price.type= request.POST.get("edit-price-type")
         price.value= request.POST.get("edit-price")
         price.save()
-        return HttpResponseRedirect(reverse("products:all-products"))
+        return HttpResponseRedirect(reverse("products:my-shop"))
     context= {"price":price}
     return render(request,"products/edit_price.html",context)
 
 @login_required(login_url="user:login")
-@is_admin
+@is_vendor
 def delete_price(request,pk):
+    vendor= Vendor.objects.get(id=request.user.id)
     price= get_price(pk)
+    if vendor != price.product.vendor:
+        return HttpResponse("You are not allowed to access this page")
     if request.method == "POST":
         price.delete()
-        return HttpResponseRedirect(reverse("products:all-products"))
+        return HttpResponseRedirect(reverse("products:my-shop"))
     context= {"obj":price}
     return render(request,"delete.html",context)
 
@@ -149,8 +173,9 @@ def products(request):
 @is_customer
 def product_details(request,pk):
     product= get_product(pk)
-    product_contents= get_product_catalogue().get(product)[0]
-    product_prices= get_product_catalogue().get(product)[1]
+    product_details= get_product_catalogue([product])[product]
+    product_contents= product_details[0]
+    product_prices= product_details[1]
     selected_prices= request.POST.get("product-prices")
     selected_contents= request.POST.getlist("product-contents")
